% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMutilities.r
\name{backward_prob}
\alias{backward_prob}
\title{Computes backward probabilities}
\usage{
backward_prob(object, state.names = NULL, par = NULL)
}
\arguments{
\item{object}{fitted crm model}

\item{state.names}{if not NULL use these labels instead of strata.labels in model}

\item{par}{if not NULL use these values for matrix calculations}
}
\value{
array of backward probabilities (one for each id, state, occasion)
}
\description{
Computes backward probability sequence for a set of capture histories
}
\references{
Zucchini, W. and I.L. MacDonald. 2009. Hidden Markov Models for Time Series: An Introduction using R. Chapman and Hall, Boca Raton, FL. See page 61.
}
\author{
Jeff Laake
}
\keyword{utility}
