% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_pred_states.r
\name{boot_pred_states}
\alias{boot_pred_states}
\title{Parametric bootstrap of state predictions}
\usage{
boot_pred_states(model, nboot = 0, beta_only = FALSE, beta_reps = NULL)
}
\arguments{
\item{model}{fitted model object}

\item{nboot}{number of bootstrap replicates; if 0 it returns the model state predictions with global_decode at the estimated parameter values}

\item{beta_only}{if TRUE, generates beta_reps from model and returns it. This is for cases where the same values of beta_reps are used for another bootstrap (eg projections using survival)}

\item{beta_reps}{if NULL the values are generated; otherwise it is a matrix of values with rows being the reps}
}
\value{
If nboot=0, a matrix of model state predictions where rows are the individual histories and columns are the occasions.
If nboot>0, a list of length nboot with state prediction matrices is returned.
}
\description{
Bootstraps the state predictions for a model by assuming a multivariate normal distribution for
the model parameter estimates. For each bootstrap replicate it replaces model parameter estimates with a 
random sample from a multivariate normal where the mean vector is the fitted model parameter estimates and
the var-cov matrix is the estimated matrix from the model fit.
}
\author{
Jeff Laake
}
\keyword{utility}
