% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMutilities.r
\name{compute_matrices}
\alias{compute_matrices}
\title{Compute HMM matrices}
\usage{
compute_matrices(object, state.names = NULL, par = NULL)
}
\arguments{
\item{object}{fitted crm model}

\item{state.names}{if not NULL use these labels instead of strata.labels in model}

\item{par}{if not NULL use these values for matrix calculations}
}
\value{
list with gamma, dmat and delta arrays
}
\description{
Computes gamma, dmat and delta (initial) matrices(arrays) and returns them in a list.
}
\author{
Jeff Laake
}
\keyword{utility}
