% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{load_lib}
\alias{load_lib}
\title{Load library for TMB model}
\usage{
load_lib(model)
}
\arguments{
\item{model}{type of crm model}
}
\description{
Loads the model specific library (eg dll) for a model
}
\keyword{utility}
