% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMutilities.r
\name{local_decode}
\alias{local_decode}
\title{Local decoding of HMM}
\usage{
local_decode(object, state.names = NULL, par = NULL)
}
\arguments{
\item{object}{fitted crm model (must be an HMM model)}

\item{state.names}{names for states if over-riding defaults}

\item{par}{if not NULL use these values for matrix calculations}
}
\value{
matrix of state predictions
}
\description{
Computes state predictions one at a time for each occasion for each individual
}
\references{
Zucchini, W. and I.L. MacDonald. 2009. Hidden Markov Models for Time Series: An Introduction using R. Chapman and Hall, Boca Raton, FL. See page 80.
}
\author{
Jeff Laake
}
\keyword{utility}
