% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markedTMB-package.R
\docType{package}
\name{markedTMB-package}
\alias{markedTMB}
\alias{markedTMB-package}
\title{markedTMB: Mark-Recapture Analysis for Survival Estimation}
\description{
Functions for fitting various models to capture-recapture data including mixed-effects multistate models, multi-variate state models and multistate live and dead recovery structure for survival estimation. All models are implemented with TMB.
}
\keyword{internal}
