% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_pred_states.r
\name{Values}
\alias{Values}
\title{Summarize statistic values across parametric bootstraps of state predictions}
\usage{
Values(pp, bs, f, index = NULL)
}
\arguments{
\item{pp}{state prediction matrix at fitted model estimates}

\item{bs}{list of bootstrap replicates of state prediction matrices}

\item{f}{function to compute some desired value for each state prediction matrix}

\item{index}{if NULL, f only produces a single value; otherwise index specifies which of the values should be summarized}
}
\value{
A dataframe containing the estimate, std error, 95% normal confidence interval and 95% percentile intervals.
}
\description{
From bootstraps replicates of the state predictions, compute the value of a statistic summarizing some
attribute. Computes the statistic value for the fitted model, standard error of the statistic value (std dev across bootstrap replicates),
lower and upper 95% confidence intervals assuming an asymptotic normal distribution and percentile 
confidence intervals across replicates.
}
\author{
Jeff Laake
}
\keyword{utility}
